Parse automatic tags.
During the parse phase when the grammar is generated, if the
AUTOMATIC TAGS keyword is declared, a callback is added to parse
of constructed types. This callback adds sequential tags to the
elements of the constructed type (unless a tag exists, as per the
spec). The callback is not added when the module tag default is
IMPLICIT or EXPLICIT.

During creation of the semantic model, if the default tag type is
AUTOMATIC, the tagged element is tagged IMPLICIT unless it is a
CHOICE type or resolves to one.

Minor bugfix: if default tag mode is IMPLICIT, CHOICEs must still
be tagged EXPLICIT.

Resolves #20