Test runner now supports multi-modules
Instead of piping text to python, we now write out all modules in an
.asn file to individual Python scripts under _testdir/. The basic_test
scripts create and remove the test directory, and check that all files
in the directory are valid Python modules.

This uncovered a couple of bugs with value references, so temporarily
disable these tests.

Note basic_test.bat is untested, I no longer have a Windows rig for
testing.