Make SingleValueConstraint accept alternatives
SingleValueConstraint can describe a set of valid values, not
necessarily just one.

This is not strictly correct -- the "|" token is used to create a UNION
constraint composed of multiple other constraints of different types, so
e.g. (1 | a | 6..10) is valid ASN.1, but not accepted by asn1ate.

When proper combined constraints are implemented, this should fall out
nicely and just work, but for now, support a common scenario with more
than one SingleValue.